Pokeball completed
The estimated total score is 100 points. 15 points were earned for the complexity of the chosen image and 15 points for its resemblance to the original. Since multiple HTML elements were used, the maximum 50 points were awarded in that category. Additionally, 10 points were earned for implementing ::before and ::after, and another 10 points for effectively using display: flex.